Selected Team Winning Logic, UI/UX - NBA
NFL -

HTML - added scoreboard api as a script element

JS - added logic to account for "scheduled" games, or games that have not been played yet in the most recent time period

NBA -

JS - Updated to match the NFL page's functionality. Added in functionality to parse out Time in EST for Scheduled Games

Bug 1: I was not accounting for scheduled games. Add in logic to return a different statement to the end user

Bug 2: Script is not working for when a team does not have a game scheduled. This is different from when a team is classified to have a BYE week in the NFL

HTML - Updated HTML to match the NFL's pages functionality

CSS - Updated CSS to match the NFL's pages functionality